Prevent Service Principal password from being reset to "" for updates
This change fixes a bug that was caused when there was an update applied
to a service connection. The bug would result in the service principal
secret being reset to the empty string.

This change fixes the problem by catching this case and setting the
value to "null", which indicates to the AzDO services that the
password should not be updated.